requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
transformers==4.36.0
torch==2.1.0